Tackle the endless global flags pollution problem
It was starting to get a bit out of control in terms of arguments in docs / help as we added new args, especially those that will only likely be used by internal Ably developers. This commit fixes that by hiding the arguments which will only likely be used by Ably developers from help and README docs.